+++ src/index.ts
@@ -8,7 +8,6 @@ export { useLoading } from './api/useLoading';
 export { getLoading } from './api/getLoading';
 export { connect } from './redux/connect';
 export { useComputed } from './reactive/useComputed';
-export { useDefined } from './model/useDefined';
 
 // 入口使用
 export { compose } from 'redux';
@@ -27,6 +26,6 @@ export type {
   StoreEnhancer,
   Unsubscribe,
 } from 'redux';
-export type { Model, HookModel } from './model/types';
+export type { Model } from './model/types';
 export type { StorageEngine } from './engines';
 export type { ComputedRef } from './reactive/types';