+++ src/model/defineModel.ts
@@ -206,7 +206,7 @@ export const defineModel = <
   }
 
   if (events) {
-    const { onInit, onChange, onDestroy } = events;
+    const { onInit, onChange } = events;
     const eventCtx: EventCtx<State> = Object.assign(
       composeGetter({ name: uniqueName }, getState),
       enhancedMethods.external,
@@ -229,19 +229,6 @@ export const defineModel = <
               onChange.call(eventCtx, prevState, nextState);
             }
             prevState = nextState;
-          }),
-        );
-      }
-
-      if (onDestroy) {
-        subscriptions.push(
-          modelStore.subscribe(() => {
-            if (eventCtx.state === void 0) {
-              for (let i = 0; i < subscriptions.length; ++i) {
-                subscriptions[i]!();
-              }
-              onDestroy.call(null as never);
-            }
           }),
         );
       }